aucode added

+++ entrance.py
@@ -1,4 +1,9 @@
 #-*- coding=utf-8 -*-
+
+aucode="cjsoft"
+maxuploadlimit=20*1024*1024
+
+
 import sys,getopt,shutil
 import os,flask,json,cjs
 import uuid
@@ -109,17 +114,24 @@ def mfr(formname=None):
     except BaseException,e:
         print e
         return flask.render_template("template.html",title="CJSoft Info Collector/%s"%formname,content=("<h2>%s</h2><a href=/form style=\"font-size:18px\">返回</a><br><br>唔，遇到了一点错误,这极有可能是表单编写不规范造成的"%formname).decode("utf-8"))
-@app.route("/export/<dbname>")
+@app.route("/export/<dbname>",methods=["POST","GET"])
 def exportresults(dbname):
-    if(os.path.isfile(getpath(os.path.join("infoforms",dbname,"db.lock")))):
-        sqlpath=cjs.export(dbname)
-        rstr=export.collect_uploads(dbname,sqlpath)
-        psplit=os.path.split(rstr)
-        f= flask.send_from_directory(psplit[0],psplit[1])
-        os.remove(rstr)
-        return f
+    if request.method=="GET":
+        return flask.render_template("template.html",title="CJSoft Info Collector",content=flask.render_template("au.html",fn=dbname,extra=""))
     else:
-        return flask.render_template("template.html",title="CJSoft Info Collector",content=("<h2>导出失败，没有可供导出的数据</h2>").decode("utf8"))
+        global aucode
+        if request.form.to_dict()["aucode"]==aucode:
+            if(os.path.isfile(getpath(os.path.join("infoforms",dbname,"db.lock")))):
+                sqlpath=cjs.export(dbname)
+                rstr=export.collect_uploads(dbname,sqlpath)
+                psplit=os.path.split(rstr)
+                f= flask.send_from_directory(psplit[0],psplit[1])
+                os.remove(rstr)
+                return f
+            else:
+                return flask.render_template("template.html",title="CJSoft Info Collector",content=("<h2>导出失败，没有可供导出的数据</h2>").decode("utf8"))
+        else:
+            return flask.render_template("template.html",title="CJSoft Info Collector",content=flask.render_template("au.html",fn=dbname,extra="鉴权码错误".decode("utf8")))
 @app.route("/favicon.ico")
 def favicon():
     return app.send_static_file("favicon.ico")
@@ -135,7 +147,7 @@ for op,value in opts:
 reload(sys)
 #cjs.connect()
 sys.setdefaultencoding('utf-8')
-app.config['MAX_CONTENT_LENGTH'] = 20 * 1024 * 1024
+app.config['MAX_CONTENT_LENGTH'] = maxuploadlimit
 
 app.debug=True
 app.run(host=addres,port=por)
